refactor: [#116] config 파일 분리
