+++ back/config/database.config.ts
@@ -1,0 +1,19 @@
+import * as dotenv from 'dotenv';
+import { TypeOrmModuleOptions } from '@nestjs/typeorm';
+
+dotenv.config();
+
+const databaseConfig: TypeOrmModuleOptions = {
+  type: 'mysql',
+  host: process.env.DB_HOST,
+  username: process.env.DB_USERNAME,
+  port: +process.env.DB_PORT,
+  password: process.env.DB_PASSWORD,
+  database: process.env.DB_DATABASE,
+  entities: [__dirname + '/../**/*.entity{.ts,.js}'],
+  // autoLoadEntities: true,
+  synchronize: false,
+  logging: true,
+};
+
+export default databaseConfig;
